move direction module up a level
